requirements.txt
asgiref==2.3.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
Babel==2.6.0
channels==2.1.7
constantly==15.1.0
daphne==2.3.0
Django==2.1.7
django-braces==1.13.0
django-crispy-forms==1.7.2
django-model-utils==3.1.2
django-phonenumber-field==2.2.0
django-private-chat==0.2.2
filelock==3.0.10
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
phonenumbers==8.10.7
Pillow==5.4.1
pkg-resources==0.0.0
pluggy==0.9.0
py==1.8.0
PyHamcrest==1.9.0
pytz==2018.9
six==1.12.0
toml==0.10.0
tox==3.8.4
Twisted==18.9.0
txaio==18.8.1
virtualenv==16.4.3
websockets==3.2
zope.interface==4.6.0
